requirements.txt
spotipy
psycopg2
django
django-environ
